+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+- `compare` method to score predictions relative to each other
+
 ## [0.2.0] - 2020-05-15
 
 ### Added